py/googleDrive/requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
oauth2client